% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{get_age_for_clade}
\alias{get_age_for_clade}
\title{Extracts the age of a given clade group from the tree.}
\usage{
get_age_for_clade(tree, clade, ages = NULL)
}
\arguments{
\item{tree}{the phylogeny}

\item{clade}{the clade label}

\item{ages}{An age dataframe from `get_ages` (define this to speed up repeated evaluation)}
}
\value{
float.
}
\description{
Extracts the age of a given clade group from the tree.
}
\examples{
get_age_for_clade(ape::rtree(5), c('t1', 't3'))
}
