% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{get_ages}
\alias{get_ages}
\title{Returns a table of tips and nodes in the tree and their relevant ages.}
\usage{
get_ages(tree)
}
\arguments{
\item{tree}{the phylogeny}
}
\value{
A data frame (as tibble) of Node and Tip Ages.
}
\description{
The columns are:
  - parental.node - the parent node
  - daughter.node - the daughter node
  - dist.root - the distance from the root
  - BL - this branch length (mrca.age - dist.tip)
  - dist.tip - the distance of the tip.
  - mrca.age - age of the most recent common ancestor (BL + dist.tip)
}
\note{
taken from https://grokbase.com/t/r/r-sig-phylo/116m5s3fr4/r-nodes-and-taxa-depth
}
\examples{
get_ages(ape::rtree(4))
}
