% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{get_nodeages}
\alias{get_nodeages}
\title{Returns a tibble of the node ages for the given tree.}
\usage{
get_nodeages(tree)
}
\arguments{
\item{tree}{the phylogeny}
}
\value{
A data frame (as tibble) of Node Ages.
}
\description{
Returns a tibble of the node ages for the given tree.
}
\examples{
tree <- ape::read.tree(text = "((t4:1,t5:2):3,((t3:4,t1:5):6,t2:7):8);")
get_nodeages(tree)
}
