% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{get_non_languages}
\alias{get_non_languages}
\title{Functions to operate on language data
Returns a dataframe of taxa to remove}
\usage{
get_non_languages(taxa)
}
\arguments{
\item{taxa}{A dataframe with columns ID, Name, Glottocode, Level}
}
\value{
A dataframe of (ID, Name, Glottocode, Reason)
}
\description{
Dialect handling:
  - go over dialects,
  - find the ID of the parent language of each dialect (=Language_ID)
  - if we do not have any other entries in the taxa table for that Language_ID
    (i.e. no other samples of that language), we 'promote' the dialect to a
    language.
  - otherwise we remove dialects
}
\examples{
taxa <- data.frame(
    ID=c("abaga", "aeka"),
    Name=c("abaga", "aeka"),
    Glottocode=c("abag1245", "aeka1244"),
    Level=c("language", "dialect"),
    Language_ID=c("", "musa1266"))
get_non_languages(taxa)
}
