% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rootheight.R
\name{get_rootheight}
\alias{get_rootheight}
\title{Returns the rootheight of a tree}
\usage{
get_rootheight(tree)
}
\arguments{
\item{tree}{the phylogeny (ape::phylo or treeio::treedata)}
}
\value{
a numeric
}
\description{
Returns the rootheight of a tree
}
\examples{
get_rootheight(ape::rcoal(4))
}
