% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{process_trees}
\alias{process_trees}
\title{Extracts a dataframe (as tibble) for all the clades from a given tree (or list
of trees).}
\usage{
process_trees(clades, trees, verbose = FALSE)
}
\arguments{
\item{clades}{a list of clades}

\item{trees}{the phylogeny}

\item{verbose}{show progress messages or not}
}
\value{
dataframe(tibble)
}
\description{
Extracts a dataframe (as tibble) for all the clades from a given tree (or list
of trees).
}
\examples{
tree <- ape::rtree(5)
clades <- list(A = c('t1', 't3'), B = c('t1', 't2', 't3'))
process_trees(clades, tree)
}
