% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{read_clade_file}
\alias{read_clade_file}
\title{Reads a clade file and parses it.
The format is:
   taxset t1t3 = t1 t3;
   taxset t4t5 = t4 t5;}
\usage{
read_clade_file(filename)
}
\arguments{
\item{filename}{the clade file to read}
}
\description{
Reads a clade file and parses it.
The format is:
   taxset t1t3 = t1 t3;
   taxset t4t5 = t4 t5;
}
