% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_tree.R
\name{rescale_tree}
\alias{rescale_tree}
\title{Rescales a tree}
\usage{
rescale_tree(tree, scaler)
}
\arguments{
\item{tree}{the phylogeny (treeio:treedata, ape::phylo, or ape::multiPhylo)}

\item{scaler}{a numeric value to rescale tree by}
}
\value{
a phylogeny or a tree
}
\description{
Rescales a tree
}
\examples{
tree <- rescale_tree(ape::rtree(4), scaler=10)
}
