% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{get_age_for_clade}
\alias{get_age_for_clade}
\title{Extracts the age of a given clade group from the tree.}
\usage{
get_age_for_clade(tree, ages, clade)
}
\arguments{
\item{tree}{the phylogeny}

\item{clade}{the clade label}

\item{age}{An age dataframe from `get_ages`}
}
\value{
float.
}
\description{
Extracts the age of a given clade group from the tree.
}
\examples{
tree <- ape::rtree(5)
ages <- get_ages(tree)
get_age_for_clade(tree, ages, c('t1', 't3'))
}
