% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{get_ages}
\alias{get_ages}
\title{Returns a table of tips and nodes in the tree and their relevant ages.
The columns are:
   parental.node - the parent node
   daughter.node - the daughter node
   dist.root - the distance from the root
   BL - this branch length
   dist.tip - the distance of the tip.
   mrca.age - age of the most recent common ancestor}
\usage{
get_ages(tree)
}
\arguments{
\item{tree}{the phylogeny}
}
\value{
A data frame (as tibble) of Node and Tip Ages.
}
\description{
Returns a table of tips and nodes in the tree and their relevant ages.
The columns are:
   parental.node - the parent node
   daughter.node - the daughter node
   dist.root - the distance from the root
   BL - this branch length
   dist.tip - the distance of the tip.
   mrca.age - age of the most recent common ancestor
}
\note{
taken from https://grokbase.com/t/r/r-sig-phylo/116m5s3fr4/r-nodes-and-taxa-depth
}
\examples{
get_ages(ape::rtree(4))
}
